+++ NEWS.md
@@ -1,3 +1,5 @@
+# fusen (development version)
+
 # fusen 0.4.0
 
 ## New features