+++ NEWS.md
@@ -1,9 +1,13 @@
-# fusen (development version)
+# fusen 0.4.1
 
 ## New features
 
 - Load all `function` chunks of the flat file currently opened with `load_flat_functions()` (Like a `load_all()` for a flat file)
 - Allow to `inflate()` a Quarto ".qmd" flat file (#160)
+
+## Minor
+
+- Fix HTML5 doc
 
 # fusen 0.4.0